JPF: Change the usage of the command manager so that ConnectionManager creates an instance that is then reused by all connections, rather than one instance per connection.
Move the load/get methods for ftpcommands.conf from FtpConfig into ConnectionManager.

Introduce a Session object which is a subclass of KeyedMap and the superclass of BaseFtpConnection, pass this into CommandRequest instead of a BaseFtpConnection object, this allows access to the command properties and provides a place for commands to store per connection data.

A few other miscellaneous tidyups.

git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/branches/jpf@1664 c83153ee-f3e8-0310-b165-dc3a72e09794