Port r1683 from trunk:
java.io.File.list() may return null when the directory no longer exists or when
an I/O error occurs.  In either case, we have successfully deleted the file
already so we can return as usual.  This change makes sure we do not use a null
pointer when list() returns null.

Thanks to Dom for reporting this issue and helping to provide a root cause and
potential fixes.  Also thanks to zubov for his useful input.

git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/branches/2.0-STABLE@1702 c83153ee-f3e8-0310-b165-dc3a72e09794