TRUNK:
The console version of the installer will now take account of the available terminal size and use all available screen space, it does however only note the size at launch, so resizing the terminal once the installer is running will not work. I cannot immediately see a way in the API to trap the terminal being resized, unless I can find a way of doing this the limitation will remain.

git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/trunk@1886 c83153ee-f3e8-0310-b165-dc3a72e09794