Change all the ant build scripts to use commonly defined macros, this makes the individual plugin build.xml files very simple which should make life easier for third party developers, additionally it will make future changes to the build far less tedious.
A few changes were required in the custom tasks to allow them to run even if the plugin doesn't need them, whilst this may add some overhead it again makes life easier for developers since they don't have to remember if they need to call them. As this is a build process some extra overhead isn't really an issue.

git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/trunk@1907 c83153ee-f3e8-0310-b165-dc3a72e09794