Upgrade the EventBus dependency to 1.3 , get the new version here https://eventbus.dev.java.net/files/documents/4303/135894/eventbus-1.3.jar and place the jar in your lib directory before build.
The use of EventBus has been changed to use their Annotations, this means we can have one onEvent() style method in subscriber classes for each event type rather than having to do nasty instanceof checks and runtime casting. Additionally the access of the singleton EventBus instance via GlobalContext has been removed in favour of the singleton support provided by the EventBus API itself, this is the method the annotated methods use to get at the instance so it is more consistant if this approach is used in all cases.

A bunch of missing svn properties have been added and some existing ones fixed.

git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/trunk@1918 c83153ee-f3e8-0310-b165-dc3a72e09794