Fix an issue with updating parent directory sizes which was introduced in [2187], an effort was made in the earlier commit to reduce the number of size updates performed and this meant the old size was never removed from the parent (but the new size still added), this fix retains the reduced number of updates whilst correctly updating the parent size.
Also ensure the parent size is decremented on wipe and delete, this was broken in [2187] as well.

Add some commits when size and lastmodified are changed on inodes to ensure these are persisted properly, this was previously unsafe to do but is possible now that we know when the inode has finished deserialization.

git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/trunk@2198 c83153ee-f3e8-0310-b165-dc3a72e09794