This fixes removing links on wiped directories.  Fixes #414.  I added a WIPE method for zip files as well.  However there is an underlying issue with wiping zip files vs deleting them.  Once that is fixed this will work correctly.  I will create a ticket for this specific issue.
The problem with the wipe command was it took the current directory, so if the user was in root dir, and wiped a section + sub folder it would fail.  Will this be the same for STOR/RETR/DELE.  I know people shouldn't be doing this, but what if someone types "STOR /section1/dir1/file.ext" will it fail, or try to upload it.

git-svn-id: https://drftpd.svn.sourceforge.net/svnroot/drftpd/trunk@2307 c83153ee-f3e8-0310-b165-dc3a72e09794